Move all the multi line tests to the new rule class to ensure the rule is adapted
